// src/app/pages/registradoras/components/registradoras-table/registradoras.interface.ts
export interface RegistradorasInterface {
   idregistradora?: number;
   montoIngresa?: number;
   montoEgresa?: number;
   montoPagando?: number;
   fecha?: string;
   hora?: string;
   empleado_idempleado?: number;
   ticket_idticket?: number;
   baja?: boolean;
   modified_at?: string;
   created_by?: number;
   created_at?: string;
}